+++ tests/Config/Loader/Parser/YamlParserTest.php
@@ -9,14 +9,24 @@ class YamlParserTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @test
+     */
+    public function construct()
+    {
+        $parser = new YamlParser(true, true, true);
+        $expected = new \stdClass();
+        $expected->yaml = 'valid';
+        self::assertEquals($expected, $parser->parse('yaml: valid'));
+    }
+
+    /**
+     * @test
      * @dataProvider parserProvider
      *
      * @param ParserInterface $parser
      */
     public function parseSuccess(ParserInterface $parser)
     {
-        $content = 'yaml: valid';
-        $result = $parser->parse($content);
+        $result = $parser->parse('yaml: valid');
         self::assertArrayHasKey('yaml', $result);
     }
 
@@ -29,8 +39,7 @@ class YamlParserTest extends \PHPUnit_Framework_TestCase
      */
     public function parseFail(ParserInterface $parser)
     {
-        $content = 'yaml: { invalid }: true';
-        $parser->parse($content);
+        $parser->parse('yaml: { invalid }: true');
     }
 
     /**